kik 15.8

+++ Kik/unknown/com/kik/kin/P2PPayment.kt
@@ -4,10 +4,11 @@ import com.kik.core.network.xmpp.jid.BareJid
 import kik.core.chat.profile.KinUserId
 import kik.core.kin.PaymentMetaData
 import kik.core.kin.PaymentType
+import java.math.BigDecimal
 import java.util.UUID
 
 
-data class P2PPayment(val recipient: BareJid, val kinUserId: KinUserId, val amount: Int, val type: PaymentType, val metaData: PaymentMetaData, val id: UUID = UUID.randomUUID()) {
+data class P2PPayment(val recipient: BareJid, val kinUserId: KinUserId, val amount: BigDecimal, val type: PaymentType, val metaData: PaymentMetaData, val id: UUID = UUID.randomUUID()) {
     var paymentJwt: String = ""
     var confirmationJwt: String = ""
 }
