+++ Kik/unknown/kik/android/chat/vm/tipping/IGroupTippingInputViewModel.kt
@@ -3,6 +3,7 @@ package kik.android.chat.vm.tipping
 import kik.android.chat.view.AbstractValidateableInputView
 import kik.android.chat.vm.IViewModel
 import rx.Observable
+import java.math.BigDecimal
 
 /**
  * Created by Simran Kaur on 2018-09-23.
@@ -18,17 +19,17 @@ interface IGroupTippingInputViewModel: IViewModel {
     /**
      * Current amount tipped as integer value
      */
-    fun getAmountTipped(): Int
+    fun getAmountTipped(): BigDecimal
 
     /**
      * Current Kin balance as observable integer
      */
-    val kinBalance: Observable<Int>
+    val kinBalance: Observable<BigDecimal>
 
     /**
      * Maximum value slider can reach (affected by daily spending limits)
      */
-    val maxValue: Observable<Int>
+    val maxValue: Observable<BigDecimal>
 
     /**
      * Object that validates input and returns an observable boolean
@@ -55,5 +56,5 @@ interface IGroupTippingInputViewModel: IViewModel {
     /**
      * Daily tipping limit remaining
      */
-    val limitRemaining: Observable<Double>
+    val limitRemaining: Observable<BigDecimal>
 }
